\name{estSimpson}
\alias{estSimpson}

\title{ Simpson index }
\description{
 Calculates the estimate of the Simpson index from a vector of counts.
}
\usage{
estSimpson(x)
}

\arguments{
  \item{x}{a integer (numeric) vector, giving the counts of p species in a community }
}


\keyword{ htest }
\concept{ biodiversity }
