\name{print.binomORci}
\alias{print.binomORci}
\title{ Print method for binomORci}
\description{
Print method for objects of class "binomORci" as can be created by calling binomORci. 
  }

\usage{
\method{print}{binomORci}(x, ...)
}

\arguments{
  \item{x}{ an object of class "binomRDci"  }
  \item{\dots}{ ... }
}

\value{
A print out.
}


\keyword{ print }

