\name{print.binomRDci}
\alias{print.binomRDci}
\title{ Print method for binomRDci }
\description{Print method for objects of class "binomRDci" as can be created by calling binomRDci. 
  }
\usage{
\method{print}{binomRDci}(x, digits = 4, ...)
}
\arguments{
  \item{x}{ an object of class "binomRDci" }
  \item{digits}{ a single integer value specifying the number of digits for rounding }
  \item{\dots}{ further arguments to be passed to "print.default" }
}

\value{
A print out.  
}


\keyword{ print }