\name{print.binomRDtest}
\alias{print.binomRDtest}
\title{ Print method for binomRDtest }
\description{
  Print method for objects of class "binomRDtest" as can be created by calling binomRDtest. 

}
\usage{
\method{print}{binomRDtest}(x, digits = 4, ...)
}
\arguments{
  \item{x}{ an object of class "binomRDtest" }
  \item{digits}{ a single integer value specifying the number of digits for rounding }
  \item{\dots}{  further arguments to be passed to "print.default" }
}
\value{
 A print out.
}


\keyword{ print }