\name{print.binomRRci}
\alias{print.binomRRci}
\title{ Print method for binomRRci }
\description{Print method for objects of class "binomRRci" as can be created by calling binomRDci. 
  }
\usage{
\method{print}{binomRRci}(x, digits = 4, ...)
}
\arguments{
  \item{x}{ an object of class "binomRRci" }
  \item{digits}{ a single integer value specifying the number of digits for rounding }
  \item{\dots}{ further arguments to be passed to "print.default" }
}

\value{
A print out.  
}


\keyword{ print }