\name{print.poly3ci}
\alias{print.poly3ci}
\title{ Print method for poly3ci }
\description{Print method for objects of class "poly3ci" as can be created by calling poly3ci 
  }
\usage{
\method{print}{poly3ci}(x, digits = 4, ...)
}
\arguments{
  \item{x}{ an object of class "poly3ci" }
  \item{digits}{ a single integer value specifying the number of digits for rounding }
  \item{\dots}{  further arguments to be passed to "print.default" }
}

\value{
A print out.  
}


\keyword{ print }