\name{print.poly3est}
\alias{print.poly3est}
\title{ Print function for poly3est }
\description{ Print function for objects of class "poly3est", as can be created by calling "poly3estf"}
\usage{

\method{print}{poly3est}(x, digits = 4, ...)
}
\arguments{
  \item{x}{ An object of class "poly3est", as can be created by calling \link{poly3estf} }
  \item{digits}{ a single integer, the number of digits for rounding of the results }
  \item{\dots}{  currently not used... }
}


\keyword{ print }