\name{print.poly3test}
\alias{print.poly3test}
\title{ Print method for poly3test }
\description{
  Print method for objects of class "poly3test" as can be created by calling poly3test. 

}
\usage{
\method{print}{poly3test}(x, digits = 4, ...)
}
\arguments{
  \item{x}{ an object of class "poly3test" }
  \item{digits}{ a single integer value specifying the number of digits for rounding }
  \item{\dots}{  further arguments to be passed to "print.default" }

}
\value{
 A print out.
}



\keyword{ print }