\name{print.Simpsonci}
\alias{print.Simpsonci}

\title{ Print function for Simpsonci }
\description{
 Print out the confidence intervals calculated by Simpsonci.
}
\usage{
\method{print}{Simpsonci}(x, ...)
}

\arguments{
  \item{x}{ An object of class \code{"Simpsonci"}, as produced by calling \code{link{Simpsonci}}}
  \item{\dots}{ currently only \code{digits} is passed to \code{print} }
}

\value{
A print out.
}
\keyword{ print }