\name{printfunctions}
\alias{print.binomORci}
\alias{print.binomRDci}
\alias{print.binomRDtest}
\alias{print.binomRRci}
\alias{print.poly3ci}
\alias{print.poly3est}
\alias{print.poly3test}
\alias{print.Shannonci}
\alias{print.Simpsonci}


\title{Print methods for the classes in this package}
\description{
Print methods for objects of the corresponding classes.
  }

\usage{
\method{print}{binomORci}(x, ...)
\method{print}{binomRDci}(x, digits = 4, ...)
\method{print}{binomRDtest}(x, digits = 4, ...)
\method{print}{binomRRci}(x, digits = 4, ...)
\method{print}{poly3ci}(x, digits = 4, ...)
\method{print}{poly3est}(x, digits = 4, ...)
\method{print}{poly3test}(x, digits = 4, ...)
\method{print}{Shannonci}(x, ...)
\method{print}{Simpsonci}(x, ...)
}

\arguments{
  \item{x}{ an object of the corresponding class }
  \item{digits}{the number of digits to be used for rounding }
  \item{\dots}{ ... }
}

\value{
A print out.
}


\keyword{ print }

