\name{print.Simpsonci}
\alias{print.Simpsonci}

\title{ Print function for Simpsonci }
\description{
 Print out the confidence intervals calculated by Simpsonci.
}
\usage{
\method{print}{Simpsonci}(x, ...)
}

\arguments{
  \item{x}{ An object of class \kbd{"Simpsonci"}, as produced by calling \code{\link{Simpsonci}}}
  \item{\dots}{ currently only \kbd{digits} is passed to \kbd{print} }
}

\value{
A print out.
}
\keyword{ print }